Implictly cast masked frame to uint8

+++ util/image.py
@@ -72,8 +72,7 @@ class ImageDiff(WindowClass):
     def __mask(fframe, fprevframe, fill_value):
         # Mask frame over old frame
         # If element is different, change value to fill_value
-        masked_frame = np.where((fframe != fprevframe).any(axis=2, keepdims=True), [fill_value,fill_value,fill_value], fframe)
-        masked_frame = masked_frame.astype(np.uint8)
+        masked_frame = np.uint8(np.where((fframe != fprevframe).any(axis=2, keepdims=True), [fill_value,fill_value,fill_value], fframe))
         return masked_frame
 
     def __frame_input(self):
